CHUKWA-211. Manage symlink correctly for RPM upgrade and uninstall. (Eric Yang)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/chukwa/trunk@775011 13f79535-47bb-0310-9956-ffa450edef68

+++ tools/specs/chukwa.spec
@@ -202,7 +202,9 @@ echo "Congratulation!  You have successfully installed Chukwa."
 %preun
 echo
 %postun
-rm -f /service/chukwa-*
+if [ ! -f /etc/init.d/chukwa-collector ]; then
+  rm -f /service/chukwa-*
+fi
 %files
 %defattr(-,%{uid},%{gid})
 %{_prefix}
